refactor: 기존 이벤트 이외 추가 이벤트는 컴포넌트로 이동
