+++ src/stores/globalStore.js
@@ -12,4 +12,5 @@ export const globalStore = createStore({
     { id: 5, author: '정수연', time: '2시간 전', content: '새로 나온 영화 재미있대요. 같이 보러 갈 사람?' },
   ],
   error: null,
+  router: null,
 });
